Add a mock model, regroup the project structure, and move the reused elements into separate views.
